+++ readium/navigator/r2-navigator/src/main/java/org/readium/r2/navigator/epub/EpubNavigatorFragment.kt
@@ -16,10 +16,10 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.fragment.app.FragmentFactory
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.viewpager.widget.ViewPager
 import kotlinx.coroutines.*
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
 import org.readium.r2.navigator.*
 import org.readium.r2.navigator.R
 import org.readium.r2.navigator.extensions.layoutDirectionIsRTL
@@ -27,8 +27,7 @@ import org.readium.r2.navigator.extensions.positionsByResource
 import org.readium.r2.navigator.pager.R2EpubPageFragment
 import org.readium.r2.navigator.pager.R2PagerAdapter
 import org.readium.r2.navigator.pager.R2ViewPager
-import org.readium.r2.navigator.pdf.PdfNavigatorFragment
-import org.readium.r2.navigator.util.SingleFragmentFactory
+import org.readium.r2.navigator.util.createFragmentFactory
 import org.readium.r2.shared.*
 import org.readium.r2.shared.publication.*
 import org.readium.r2.shared.publication.Link
@@ -36,13 +35,16 @@ import org.readium.r2.shared.publication.Locator
 import org.readium.r2.shared.publication.Publication
 import org.readium.r2.shared.publication.epub.EpubLayout
 import org.readium.r2.shared.publication.presentation.presentation
+import org.readium.r2.shared.publication.services.isRestricted
 import org.readium.r2.shared.publication.services.positions
-import kotlin.coroutines.CoroutineContext
 import kotlin.math.ceil
 
 /**
  * Navigator for EPUB publications.
+ *
+ * To use this [Fragment], create a factory with `EpubNavigatorFragment.createFactory()`.
  */
+@OptIn(ExperimentalCoroutinesApi::class)
 class EpubNavigatorFragment private constructor(
     internal val publication: Publication,
     private val baseUrl: String,
@@ -52,25 +54,8 @@ class EpubNavigatorFragment private constructor(
 
     interface Listener: VisualNavigator.Listener
 
-    /**
-     * Factory for [EpubNavigatorFragment].
-     *
-     * @param publication EPUB publication to render in the navigator.
-     * @param baseUrl A base URL where this publication is served from.
-     * @param initialLocator The first location which should be visible when rendering the
-     *        publication. Can be used to restore the last reading location.
-     * @param listener Optional listener to implement to observe events, such as user taps.
-     */
-    class Factory(
-        private val publication: Publication,
-        private val baseUrl: String,
-        private val initialLocator: Locator? = null,
-        private val listener: Listener? = null
-    ) : SingleFragmentFactory<EpubNavigatorFragment>() {
-
-        override fun instantiate(): EpubNavigatorFragment =
-            EpubNavigatorFragment(publication, baseUrl, initialLocator, listener)
-
+    init {
+        require(!publication.isRestricted) { "The provided publication is restricted. Check that any DRM was properly unlocked using a Content Protection."}
     }
 
     internal lateinit var positions: List<Locator>
@@ -103,7 +88,7 @@ class EpubNavigatorFragment private constructor(
         resourcesDouble = ArrayList()
 
         positions = runBlocking { publication.positions() }
-        publicationIdentifier = publication.metadata.identifier!!
+        publicationIdentifier = publication.metadata.identifier ?: publication.metadata.title
 
         val supportFragmentManager = currentActivity.supportFragmentManager
 
@@ -399,8 +384,8 @@ class EpubNavigatorFragment private constructor(
     override val readingProgression: ReadingProgression
         get() = publication.contentLayout.readingProgression
 
-    override val currentLocator: LiveData<Locator?> get() = _currentLocator
-    private val _currentLocator = MutableLiveData<Locator?>(null)
+    override val currentLocator: StateFlow<Locator> get() = _currentLocator
+    private val _currentLocator = MutableStateFlow(initialLocator ?: publication.readingOrder.first().toLocator())
 
     /**
      * While scrolling we receive a lot of new current locations, so we use a coroutine job to
@@ -422,13 +407,29 @@ class EpubNavigatorFragment private constructor(
             val locator = positions[positionIndex]
                     .copyWithLocations(progression = progression)
 
-            if (locator == currentLocator.value) {
+            if (locator == _currentLocator.value) {
                 return@launch
             }
 
-            _currentLocator.postValue(locator)
+            _currentLocator.value = locator
             navigatorDelegate?.locationDidChange(navigator = navigator, locator = locator)
         }
+    }
+
+    companion object {
+
+        /**
+         * Creates a factory for [EpubNavigatorFragment].
+         *
+         * @param publication EPUB publication to render in the navigator.
+         * @param baseUrl A base URL where this publication is served from.
+         * @param initialLocator The first location which should be visible when rendering the
+         *        publication. Can be used to restore the last reading location.
+         * @param listener Optional listener to implement to observe events, such as user taps.
+         */
+        fun createFactory(publication: Publication, baseUrl: String, initialLocator: Locator? = null, listener: Listener? = null): FragmentFactory =
+            createFragmentFactory { EpubNavigatorFragment(publication, baseUrl, initialLocator, listener) }
+
     }
 
 }