+++ readium/navigator/r2-navigator/src/main/java/org/readium/r2/navigator/pdf/R2PdfActivity.kt
@@ -13,6 +13,7 @@ import android.app.Activity
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.Observer
+import androidx.lifecycle.asLiveData
 import org.readium.r2.navigator.Navigator
 import org.readium.r2.navigator.R
 import org.readium.r2.shared.PdfSupport
@@ -31,12 +32,12 @@ abstract class R2PdfActivity : AppCompatActivity(), PdfNavigatorFragment.Listene
     /**
      * Override this event handler to save the current location in the publication in a database.
      */
-    abstract fun onCurrentLocatorChanged(locator: Locator)
+    open fun onCurrentLocatorChanged(locator: Locator) {}
 
     override fun onCreate(savedInstanceState: Bundle?) {
         publication = intent.getPublication(this)
 
-        supportFragmentManager.fragmentFactory = PdfNavigatorFragment.Factory(
+        supportFragmentManager.fragmentFactory = PdfNavigatorFragment.createFactory(
             publication = publication,
             initialLocator = intent.getParcelableExtra("locator"),
             listener = this
@@ -46,7 +47,7 @@ abstract class R2PdfActivity : AppCompatActivity(), PdfNavigatorFragment.Listene
 
         setContentView(R.layout.r2_pdf_activity)
 
-        navigator.currentLocator.observe(this, Observer { locator ->
+        navigator.currentLocator.asLiveData().observe(this, Observer { locator ->
             locator ?: return@Observer
 
             onCurrentLocatorChanged(locator)